+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change `@alanszp/access-list`: Change of logic on how we detect an admin on access list
 - Add `@alanszp/access-list`: Add method `whichEmployeesHasAccess` and `getFullAccessList` to `AccessListClient`
 - Change `@alanszp/access-list`: Add documentation
+- Removes `@alanszp/access-list`: Removes from `AccessListClient` the method `getSegmentReference` in favour of using the prop `segmentReference`
 
 ## v15.1.0