+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v16.0.0
 
+- Add `@alanszp/core`: Add generic typescript types.
 - Change `@alanszp/access-list`: Change of logic on how we detect an admin on access list
 - Add `@alanszp/access-list`: Add method `whichEmployeesHasAccess` and `getFullAccessList` to `AccessListClient`
 - Change `@alanszp/access-list`: Add documentation